moving files to a lib directory. better in line with other nodejs projects
