+++ lib/grant.js
