+++ lib/token.js
